+++ src/main/java/com/techshroom/emergencylanding/library/lwjgl/render/TextureRender.java
@@ -35,7 +35,7 @@ public class TextureRender extends Render<ELEntity> {
     public void doRender(ELEntity entity, float posX, float posY, float posZ) {
         DrawableUtils.beginStandardEntityRender(entity, posX, posY, posZ);
         VBAO quad = Shapes.getQuad(new VertexData(),
-                new VertexData().setXYZ((float) entity.getTex().getWidth(), (float) entity.getTex().getHeight(), 0),
+                new VertexData().setXYZ(entity.getTex().getWidth(), entity.getTex().getHeight(), 0),
                 Shapes.XY);
         quad.setTexture(entity.getTex());
         quad.setStatic(false);
